+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Tip
 
+* Add support for database/sql.Scanner and database/sql/driver.Valuer interfaces
 * Go float64 can no longer be encoded to a PostgreSQL float4
 * Add ConnPool.Reset method
 * []byte skips encoding/decoding